Fix annoying decimal battery level announcement

+++ bb/src/main/java/com/richardmcdougall/bb/BatterySupervisor.java
@@ -126,7 +126,7 @@ public class BatterySupervisor {
 
         }
         if (announce) {
-            service.speak("Battery Level is " + level + " percent", "batteryLow");
+            service.speak("Battery Level is " + (int)level + " percent", "batteryLow");
         }
     }
